improved ecisting routes acrt and added new one for mouse press

+++ server/src/controllers/consumer/cart/patch.ts
@@ -6,9 +6,53 @@ import { Cart, CartInstance } from "../../../models/all/Cart.js";
 import { CartItem } from "../../../models/all/CartItem.js";
 import { seq } from "../../../config/db.js";
 import { KEY_ACTION_CART } from "../../../types/all/cart.js";
-import { err404, err422, err500 } from "../../../lib/responseClient/err.js";
+import {
+  err400,
+  err404,
+  err422,
+  err500,
+} from "../../../lib/responseClient/err.js";
 import { User } from "../../../models/models.js";
 import { Book } from "../../../models/all/Book.js";
+import { Transaction } from "sequelize";
+
+const getCart = async ({
+  cart,
+  t,
+  userID,
+}: {
+  cart: CartInstance | null;
+  t: Transaction;
+  userID: string;
+}) => {
+  cart = await Cart.findOne({
+    where: {
+      userID,
+    },
+    include: [
+      {
+        model: CartItem,
+        as: "items",
+        include: [
+          {
+            model: Book,
+            as: "book",
+          },
+        ],
+      },
+    ],
+    transaction: t,
+  });
+  if (!cart) {
+    cart = await Cart.create(
+      {
+        userID,
+      },
+      { transaction: t }
+    );
+    cart.items = [];
+  }
+};
 
 export const patchCartByClick = async (req: ReqApp, res: Response) => {
   const {
@@ -22,38 +66,28 @@ export const patchCartByClick = async (req: ReqApp, res: Response) => {
   const t = await seq.transaction();
 
   try {
-    cart = await Cart.findOne({
-      where: {
-        userID,
-      },
-      include: [
-        {
-          model: CartItem,
-          as: "items",
-        },
-      ],
-      transaction: t,
-    });
-    if (!cart) {
-      cart = await Cart.create(
-        {
-          userID,
-        },
-        { transaction: t }
-      );
-      cart.items = [];
-    }
+    await getCart({ cart, t, userID: userID! });
 
-    const existingItem = (cart.items ?? []).find((el) => el.bookID === bookID);
+    const existingItem = (cart!.items ?? []).find((el) => el.bookID === bookID);
     if (!existingItem) {
       if (act !== KEY_ACTION_CART.INC_QTY_CART) {
         await t.rollback();
         return err404(res, { msg: "item not found" });
       }
 
+      const book = await Book.findByPk(bookID);
+      if (!book) {
+        await t.rollback();
+        return err404(res, { msg: "book not found" });
+      }
+      if (book.qty < 1) {
+        await t.rollback();
+        return err422(res, { msg: "not enough books" });
+      }
+
       await CartItem.create(
         {
-          cartID: cart.id,
+          cartID: cart!.id,
           bookID,
           qty: 1,
         },
@@ -67,10 +101,18 @@ export const patchCartByClick = async (req: ReqApp, res: Response) => {
           break;
 
         case KEY_ACTION_CART.INC_QTY_CART:
+          if (existingItem.book!.qty < existingItem.qty + 1) {
+            await t.rollback();
+            return err422(res, { msg: "not enough books" });
+          }
           await existingItem.increment("qty", { by: 1, transaction: t });
           break;
 
         case KEY_ACTION_CART.DEC_QTY_CART:
+          if (existingItem.qty - 1 < 0) {
+            await t.rollback();
+            return err400(res, { msg: "can not set qty less than 0" });
+          }
           if (existingItem.qty < 2)
             await existingItem.destroy({ transaction: t });
           else await existingItem.decrement("qty", { by: 1, transaction: t });
@@ -132,4 +174,64 @@ export const updateCartByInputTxt = async (req: ReqApp, res: Response) => {
   await cartItem.update({ qty });
 
   return res200(res, { msg: "cart updated" });
+};
+
+export const updateCartByMousePress = async (req: ReqApp, res: Response) => {
+  const {
+    userID,
+    body: { qty },
+  } = req;
+  const { bookID } = req.params;
+
+  let cart: CartInstance | null = null;
+
+  const t = await seq.transaction();
+
+  try {
+    await getCart({ cart, t, userID: userID! });
+
+    const existingItem = (cart!.items ?? []).find((el) => el.bookID === bookID);
+
+    const book = await Book.findByPk(bookID);
+    if (!book) {
+      await t.rollback();
+      return err404(res, { msg: "book not found" });
+    }
+    if (book.qty < qty) {
+      await t.rollback();
+      return err422(res, { msg: "not enough books" });
+    }
+
+    if (!existingItem) {
+      await CartItem.create(
+        {
+          bookID,
+          cartID: cart!.id,
+          qty,
+        },
+        {
+          transaction: t,
+        }
+      );
+    } else {
+      if (!qty) {
+        await existingItem.destroy({ transaction: t });
+        if (cart!.items!.length < 2) await cart!.destroy({ transaction: t });
+      } else {
+        await existingItem.update({ qty }, { transaction: t });
+      }
+    }
+
+    await t.commit();
+
+    return res200(res, { msg: "cart updated" });
+  } catch (err) {
+    await t.rollback();
+
+    console.log(err);
+
+    return err500(res);
+  }
+
+  return res200(res, {});
 };
